Moved definition of pollForm inital state to poll form file
